Refactor patch_config to portality/util to prevent extra init of LCC in anon import
