+++ doajtest/unit/test_background_task_status_parameterised.py
@@ -13,7 +13,8 @@ def load_cases():
 import json
 
 from doajtest.fixtures.background import save_mock_bgjob
-from doajtest.helpers import DoajTestCase, apply_test_case_config, patch_config
+from doajtest.helpers import DoajTestCase, apply_test_case_config
+from portality.util import patch_config
 from portality import constants
 from portality.bll import DOAJ
 from portality.tasks.anon_export import AnonExportBackgroundTask
