+++ doajtest/unit/test_tasks_old_data_cleanup.py
@@ -2,6 +2,7 @@ import datetime
 import time
 from typing import Iterable
 
+import portality.util
 from doajtest import helpers
 from doajtest.helpers import DoajTestCase
 from portality.dao import DomainObject
@@ -40,7 +41,7 @@ class TestOldDataCleanup(DoajTestCase):
         self.assertEqual(BackgroundJob.hit_count(es_queries.query_all()),
                          len(days_for_test))
 
-        org_config = helpers.patch_config(self.app_test, {
+        org_config = portality.util.patch_config(self.app_test, {
             "TASK_DATA_RETENTION_DAYS": {
                 "notification": 4,
                 "background_job": 7,
@@ -55,7 +56,7 @@ class TestOldDataCleanup(DoajTestCase):
         self.assertEqual(Notification.hit_count(es_queries.query_all()), 4)
         self.assertEqual(BackgroundJob.hit_count(es_queries.query_all()), 6)
 
-        helpers.patch_config(self.app_test, org_config)
+        portality.util.patch_config(self.app_test, org_config)
 
     def test_clean_all_old_data__no_obj_deleted(self):
         # prepare data
@@ -67,7 +68,7 @@ class TestOldDataCleanup(DoajTestCase):
         self.assertEqual(org_size, len(days_for_test))
         self.assertEqual(BackgroundJob.hit_count(es_queries.query_all()), 0)
 
-        org_config = helpers.patch_config(self.app_test, {
+        org_config = portality.util.patch_config(self.app_test, {
             "TASK_DATA_RETENTION_DAYS": {
                 "notification": 1000,
                 "background_job": 7,
@@ -82,4 +83,4 @@ class TestOldDataCleanup(DoajTestCase):
         self.assertEqual(Notification.hit_count(es_queries.query_all()), org_size)
         self.assertEqual(BackgroundJob.hit_count(es_queries.query_all()), 0)
 
-        helpers.patch_config(self.app_test, org_config)
+        portality.util.patch_config(self.app_test, org_config)
